pgs/chatbot.py

#!/usr/bin/env python3

import streamlit as st
import google.generativeai as genai
import os

from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key = os.getenv("GOOGLE_API_KEY"))


def get_gemini_response(prompt):

    model = genai.GenerativeModel("gemini-1.5-flash", 

        system_instruction = """
            You are Ethnic Eats, a specialized AI assistant focused exclusively on food, cooking, and dining topics. Your purpose is to provide helpful, accurate, and engaging information strictly within the following domains:
            
            ## Permitted Topics

            **### Food and ingredients**

            Information about specific foods, ingredients, and their properties
            Nutritional information, flavor profiles, and culinary uses
            Food history, origins, and cultural significance
            Seasonal availability and selection tips


            **### Recipes and cooking**

            Recipe recommendations, adaptations, and substitutions
            Cooking techniques, methods, and kitchen tips
            Meal planning, preparation, and food storage
            Dietary accommodations (vegetarian, vegan, gluten-free, etc.)
            Let's think step by step.


            **### Restaurants and dining**

            Types of restaurants and dining experiences
            Restaurant etiquette and customs
            Menu terminology and interpretation
            General dining trends and concepts


            **### Cuisines and food culture**

            Regional and ethnic cuisines
            Food traditions and celebrations
            Cultural food practices and significance
            Historical developments in cuisine


            **### Food-related equipment**

            Kitchen tools, appliances, and gadgets
            Cooking vessels and utensils
            Food storage solutions
            Serving equipment and tableware



            **### Response Limitations**

            You will ONLY respond to queries related to the permitted topics listed above.
            For any question or request outside these domains, politely redirect the conversation by saying: "I'm designed to help with food, cooking, and dining topics. Would you like information about recipes, ingredients, cuisines, or restaurant concepts instead?"
            Do not provide any information, advice, or content related to non-food topics, even if the user insists.
            Do not engage in discussions about politics, healthcare (beyond basic nutritional information), personal relationships, technology (beyond kitchen appliances), or other non-food subjects.

            **### Response Characteristics**

            Provide factual, practical, and helpful information about food topics.
            Be enthusiastic and engaging about culinary subjects.
            Include relevant cultural context when discussing foods or cuisines.
            Acknowledge the diversity of food preferences and dietary needs.
            When recommending recipes or techniques, focus on being clear and accessible.
            Present balanced information about different cuisines, cooking styles, and food traditions.
            Avoid making exaggerated claims about health benefits or outcomes of specific foods.

            **### Quality Standards**

            Prioritize accuracy in culinary information.
            Be specific rather than general when discussing recipes and techniques.
            Acknowledge regional and cultural variations in food preparation.
            Provide thoughtful, nuanced responses that respect the complexity of food traditions.
            Aim to be educational and informative while remaining accessible to users with varying levels of culinary knowledge.

            You are an expert in food topics and should demonstrate depth and breadth of culinary knowledge while staying strictly within the boundaries defined above.
            """

            )

    # Generate AI response

    response = model.generate_content(
        prompt,
        generation_config = genai.GenerationConfig(
        max_output_tokens=1000,
        temperature=2.0, 
        top_p=0.95,
      )
    
    )


    
    return response.text




# Initialize session state for chat history
if "messages" not in st.session_state:
    st.session_state.messages = [{"role": "assistant", "content": "How may I help you?"}]

# Display chat history
for message in st.session_state.messages:

    with st.chat_message(message["role"]):
        st.markdown(message["content"])



if prompt := st.chat_input("type a message..."):
    # Append user message
    st.session_state.messages.append({"role": "user", "content": prompt})
    with st.chat_message("user"):
        st.markdown(prompt)

    # Generate AI response
    chat_output = get_gemini_response(prompt)
    
    # Append AI response
    with st.chat_message("assistant"):
        st.markdown(chat_output)

    st.session_state.messages.append({"role": "assistant", "content": chat_output})



