requirements.txt
pandas
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
seaborn
requests
matplotlib
numpy
google-generativeai
streamlit-folium
streamlit-audiorec
folium
bcrypt
captcha
holehe
scikit-learn
bs4
csv